refactor(#1936): move MinglitAvatarImage to MDS, delegate to MinglitImage
Widget now lives in shared/packages/mds/core alongside MinglitImage and
delegates to it for the network-URL path: ClipOval(child: MinglitImage(…)).
The null/empty-URL fallback (CircleAvatar + icon) is unchanged.

All five app_user call sites now receive MinglitAvatarImage through the
existing minglit_kit → mds export chain — no local import needed.

Constraint: MinglitImage does not accept a custom loading placeholder;
shimmer + image_not_supported_outlined used for URL case.
Rejected: Keeping widget in app_user | dual maintenance of ResizeImage policy
Rejected: Adding custom frameBuilder param to MinglitImage | over-engineering
Confidence: high
Scope-risk: narrow
