+++ apps/app_user/lib/src/features/home/home_page.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:app_user/src/common/widgets/minglit_avatar_image.dart';
 import 'package:app_user/src/features/home/logic/home_coordinator.dart';
 import 'package:app_user/src/features/home/widgets/event_now_bar.dart';
 import 'package:app_user/src/features/home/widgets/event_now_bar_controller.dart';
